// src/api/generated/models/apiResponseSliceCollectionResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Dolink API Documentation
 * This is DoLink! 저장하고, 묶고, 바로 꺼내는 링크 관리
 * OpenAPI spec version: v1.0.0
 */
import type { SliceCollectionResponse } from './sliceCollectionResponse';

export interface ApiResponseSliceCollectionResponse {
  isSuccess?: boolean;
  code?: string;
  message?: string;
  result?: SliceCollectionResponse;
}