// src/api/generated/models/searchTasksParams.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Dolink API Documentation
 * This is DoLink! 저장하고, 묶고, 바로 꺼내는 링크 관리
 * OpenAPI spec version: v1.0.0
 */

export type SearchTasksParams = {
  /**
   * 검색어
   */
  keyword: string;
  /**
   * 페이지 번호 (0부터 시작)
   */
  page?: number;
  /**
   * 페이지 크기
   */
  size?: number;
};